feat: make the code base crypto agile
The code base switches from a crypto library using (PublicKey, SecretKey, Signature), and implemented with ed25519_dalek, to a code base using one generic type implementingthree one of three traits:
- `VerifyingKey` (corresponding to the notion of Public Key)
- `SigningKey` (corresponding to private keys)
- `Authenticator` (corresponding to a signature)

Every implementation of one of these traits links to compatible implementations of the two others, making implementers "come in" with a full-featured, consistent scheme.
The traits interface with concrete instantiations in the `node` crate, where the primary, workers, etc are instantiated with the Ed25519 signature scheme.

An additional trait `KeyPair` governs the usual notion of a Public / Private Key pair.
We leave runtime configuration of the key scheme when launching the node to a later PR.

- the Verifying key has an implementation of verify_batch that is overridden by those signature schemes that support batch verification,
- we plan to similarly support signature aggregation through default member override,
- as is usual in Rust, the traits are implemented over a NewType (a struct wrapper around the used library type), a requirement for data implementing a complex set of trait bounds.
- the interaction of this scheme with the benchmark infrastructure (fab files) is dependent on the serialization of KeyPairs. The trait + concrete implementation might migrate into a generic struct in order to represent those constraints, which for now are just in code.