Introduce block_synchronizer_handler and wire to block waiter (part 2) (#240)
feat: Introduce block_synchronizer_handler and wire to block waiter (part 2)

This commit is extending the block synchronizer handler to allow the payload synchronization in a synchronous way. Also the block_waiter is extended to use the block synchronizer handler to sync the (potentially) missing payloads before trying to retrieve them.