fix: turn off logging stderr+out to file (#454)
* fix: turn off logging stderr+out to file

Removing the log file so we can just log to the container
stdout and stderr to be picked up by Loki or other means.

* fix: fixed depends_on to the actual loki service name

* chore: updated README.md about logs

Specifically we changed the container entry.sh script to not
redirect the logs to files.  This enables log scraping directly
from the containers into log collection services like Loki.