feat(network): enable callers to specify an adapter for retry tasks
The retry scheduler uses a uniform return type `Result<(), RetryError>`
internally to avoid dealing with trait objects (boxing). The drawback to
this approach is that it is a leaky abstraction. Instead, we want the
type signature to reflect the method's idiosyncracy. We achieve this
by enabling callers to specify a custom adapter mapping the `RetryResult`
to a custom `Result<T, E>` that can be type inferred at compile-time.