fix: Multiplex messages to the BlockSynchronizer on a single channel
Today, the BlockSynchronizer receives several types of messages from the PrimaryReceiverHandler, which go on a different channel depending on their nature.
This does not scale, especially since we are about to add more messages types traveling the same component edge in #848.

We put the messages under an enum and transmit that enum along the channel, which simplified BlockSynchronizer setup and channel patterns overall.