Add instrumentation to block synchronizer
This adds instrumentation on every function that is called from top level select of the block synchronizer as well as to futures that are collected into `waiting`. This can shed some light into what is happening inside block synchronizer.

Also, it adds readable `Debug` implementation for `RequestId`, otherwise it is very hard to read in logs.